backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
faiss-cpu>=1.7.4
pymongo==4.7.2
motor==3.4.0
redis==5.0.4
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pdfplumber==0.11.0
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
anyio==4.4.0
slowapi==0.1.9
numpy==1.26.4
groq==0.9.0
